% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreagate_profiles.R
\name{aggregate_profiles}
\alias{aggregate_profiles}
\title{Aggregate Ceteris Paribus into Partial Dependency Plots}
\usage{
aggregate_profiles(x, ..., only_numerical = TRUE, groups = NULL,
  type = "partial", variables = NULL)
}
\arguments{
\item{x}{a ceteris paribus explainer produced with function `ceteris_paribus()`}

\item{...}{other explainers that shall be plotted together}

\item{only_numerical}{a logical. If TRUE then only numerical variables will be plotted. If FALSE then only categorical variables will be plotted.}

\item{groups}{a variable name that will be usef for grouping. By default 'NULL' which means that no groups shall be calculated}

\item{type}{either 'partial'/'conditional'/'accumulated' for parital dependence, conditional profiles of accumulated local effects}

\item{variables}{if not NULL then only `variables` will be presented}
}
\value{
an 'aggregated_ceteris_paribus_explainer' layer
}
\description{
Function 'aggregate_profiles' calculates aggregate of ceteris paribus profiles
}
\examples{
library("DALEX")
 \dontrun{
library("titanic")
library("randomForest")

titanic_small <- titanic_train[,c("Survived", "Pclass", "Sex", "Age",
                                   "SibSp", "Parch", "Fare", "Embarked")]
titanic_small$Survived <- factor(titanic_small$Survived)
titanic_small$Sex <- factor(titanic_small$Sex)
titanic_small$Embarked <- factor(titanic_small$Embarked)
titanic_small <- na.omit(titanic_small)
rf_model <- randomForest(Survived ~ Pclass + Sex + Age + SibSp + Parch + Fare + Embarked,
                         data = titanic_small)
explainer_rf <- explain(rf_model, data = titanic_small,
                        y = titanic_small$Survived == "1", label = "RF")

selected_passangers <- select_sample(titanic_small, n = 100)
cp_rf <- ceteris_paribus(explainer_rf, selected_passangers)
cp_rf

pdp_rf_p <- aggregate_profiles(cp_rf, variables = "Age", type = "partial")
pdp_rf_p$`_label_` <- "RF_partial"
pdp_rf_c <- aggregate_profiles(cp_rf, variables = "Age", type = "conditional")
pdp_rf_c$`_label_` <- "RF_conditional"
pdp_rf_a <- aggregate_profiles(cp_rf, variables = "Age", type = "accumulated")
pdp_rf_a$`_label_` <- "RF_accumulated"
plot(pdp_rf_p, pdp_rf_c, pdp_rf_a, color = "_label_")

pdp_rf <- aggregate_profiles(cp_rf, variables = "Age",
                             groups = "Sex")
pdp_rf
plot(cp_rf, variables = "Age", color = "grey") +
  show_observations(cp_rf, variables = "Age", color = "grey") +
  show_rugs(cp_rf, variables = "Age", color = "red") +
  show_aggreagated_profiles(pdp_rf, size = 3, color = "_label_")
}
}
