% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_dependence.R
\name{conditional_dependence}
\alias{conditional_dependence}
\alias{conditional_dependence.explainer}
\alias{conditional_dependence.default}
\alias{conditional_dependence.ceteris_paribus_explainer}
\alias{local_dependency}
\alias{conditional_dependency}
\title{Conditional Dependence Profiles}
\usage{
conditional_dependence(x, ...)

\method{conditional_dependence}{explainer}(
  x,
  variables = NULL,
  N = 500,
  variable_splits = NULL,
  grid_points = 101,
  ...,
  variable_type = "numerical"
)

\method{conditional_dependence}{default}(
  x,
  data,
  predict_function = predict,
  label = class(x)[1],
  variables = NULL,
  N = 500,
  variable_splits = NULL,
  grid_points = 101,
  ...,
  variable_type = "numerical"
)

\method{conditional_dependence}{ceteris_paribus_explainer}(x, ..., variables = NULL)

local_dependency(x, ...)

conditional_dependency(x, ...)
}
\arguments{
\item{x}{an explainer created with function \code{DALEX::explain()}, an object of the class \code{ceteris_paribus_explainer}
or a model to be explained.}

\item{...}{other parameters}

\item{variables}{names of variables for which profiles shall be calculated.
Will be passed to \code{\link{calculate_variable_split}}. If \code{NULL} then all variables from the validation data will be used.}

\item{N}{number of observations used for calculation of partial dependence profiles. By default \code{500}.}

\item{variable_splits}{named list of splits for variables, in most cases created with \code{\link{calculate_variable_split}}.
If \code{NULL} then it will be calculated based on validation data avaliable in the \code{explainer}.}

\item{grid_points}{number of points for profile. Will be passed to \code{\link{calculate_variable_split}}.}

\item{variable_type}{a character. If \code{"numerical"} then only numerical variables will be calculated.
If \code{"categorical"} then only categorical variables will be calculated.}

\item{data}{validation dataset, will be extracted from \code{x} if it's an explainer
NOTE: It is best when target variable is not present in the \code{data}}

\item{predict_function}{predict function, will be extracted from \code{x} if it's an explainer}

\item{label}{name of the model. By default it's extracted from the \code{class} attribute of the model}
}
\value{
an object of the class \code{aggregated_profile_explainer}
}
\description{
Conditional Dependence Profiles (aka Local Profiles) average localy Ceteris Paribus Profiles.
Function 'conditional_dependence' calls 'ceteris_paribus' and then 'aggregate_profiles'.
}
\details{
Find more details in the \href{https://ema.drwhy.ai/accumulatedLocalProfiles.html}{Accumulated Local Dependence Chapter}.
}
\examples{
library("DALEX")
library("ingredients")

model_titanic_glm <- glm(survived ~ gender + age + fare,
                         data = titanic_imputed, family = "binomial")

explain_titanic_glm <- explain(model_titanic_glm,
                               data = titanic_imputed[,-8],
                               y = titanic_imputed[,8],
                               verbose = FALSE)

cdp_glm <- conditional_dependence(explain_titanic_glm,
                                  N = 150, variables = c("age", "fare"))
head(cdp_glm)
plot(cdp_glm)

\donttest{
library("ranger")

model_titanic_rf <- ranger(survived ~., data = titanic_imputed, probability = TRUE)

explain_titanic_rf <- explain(model_titanic_rf,
                              data = titanic_imputed[,-8],
                              y = titanic_imputed[,8],
                              label = "ranger forest",
                              verbose = FALSE)

cdp_rf <- conditional_dependence(explain_titanic_rf, N = 200, variable_type = "numerical")
plot(cdp_rf)

cdp_rf <- conditional_dependence(explain_titanic_rf, N = 200, variable_type = "categorical")
plotD3(cdp_rf, label_margin = 100, scale_plot = TRUE)
}

}
\references{
Explanatory Model Analysis. Explore, Explain, and Examine Predictive Models. \url{https://ema.drwhy.ai/}
}
