% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\alias{describe.ceteris_paribus_explainer}
\title{Describe Ceteris Paribus Explainers.}
\usage{
describe(x, ...)

\method{describe}{ceteris_paribus_explainer}(x, ...)
}
\arguments{
\item{x}{a ceteris paribus explainer produced with function `ceteris_paribus()`}

\item{...}{other explainers that shall be plotted together}
}
\value{
an `ceteris_paribus_descriptions` object.
}
\description{
Function \code{describe.ceteris_paribus_explainer} generates text descriptions for Ceteris Paribus Profiles.
Note that currently only first observation is being described.
}
\examples{
\dontrun{
}
}
