% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspect_importance.R
\name{get_sample}
\alias{get_sample}
\title{Function for getting binary matrix}
\usage{
get_sample(n, m, method = c("default", "binom"), p = 0.05)
}
\arguments{
\item{n}{number of rows}

\item{m}{number of columns}

\item{method}{sampling method}

\item{p}{probability for binomial sampling}
}
\value{
a binary matrix
}
\description{
This function creates binary matrix.
It uses either binomial distribution or sample() function
}
\examples{
 \dontrun{
 get_sample(100,4,"binom")
}
}
