% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspect_importance.R
\name{group_variables}
\alias{group_variables}
\title{Group features into aspects}
\usage{
group_variables(x, p = 0.5)
}
\arguments{
\item{x}{dataframe with only numeric columns}

\item{p}{correlation value for cutoff point}
}
\value{
list of aspects
}
\description{
Divides correlated features into groups, called aspects
}
