% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceteris_paribus.R
\name{ceteris_paribus}
\alias{ceteris_paribus}
\alias{ceteris_paribus.explainer}
\alias{ceteris_paribus.default}
\title{Calculates Individual Variable Profiles aka Ceteris Paribus Profiles}
\usage{
ceteris_paribus(x, ...)

\method{ceteris_paribus}{explainer}(x, new_observation, y = NULL,
  variables = NULL, variable_splits = NULL, grid_points = 101, ...)

\method{ceteris_paribus}{default}(x, data, predict_function = predict,
  new_observation, y = NULL, variables = NULL,
  variable_splits = NULL, grid_points = 101, label = class(x)[1],
  ...)
}
\arguments{
\item{x}{a model to be explained, or an explainer created with function `DALEX::explain()`.}

\item{...}{other parameters}

\item{new_observation}{a new observation with columns that corresponds to variables used in the model}

\item{y}{true labels for `new_observation`. If specified then will be added to ceteris paribus plots.}

\item{variables}{names of variables for which profiles shall be calculated. Will be passed to `calculate_variable_splits()`. If NULL then all variables from the validation data will be used.}

\item{variable_splits}{named list of splits for variables, in most cases created with `calculate_variable_splits()`. If NULL then it will be calculated based on validation data avaliable in the `explainer`.}

\item{grid_points}{number of points for profile. Will be passed to `calculate_variable_splits()`.}

\item{data}{validation dataset, will be extracted from `x` if it's an explainer}

\item{predict_function}{predict function, will be extracted from `x` if it's an explainer}

\item{label}{name of the model. By default it's extracted from the 'class' attribute of the model}
}
\value{
An object of the class 'ceteris_paribus_explainer'.
It's a data frame with calculated average responses.
}
\description{
This explainer works for individual observations.
For each observation it calculates Individual Variable Profiles for selected variables.
For this reason it is also called 'Local Profile Plot'.
}
\examples{
library("DALEX2")
library("ceterisParibus2")
 \dontrun{
library("randomForest")
set.seed(59)

apartments_rf <- randomForest(m2.price ~ construction.year + surface + floor +
                                no.rooms + district, data = apartments)

explainer_rf <- explain(apartments_rf,
                        data = apartments_test[,2:6], y = apartments_test$m2.price)

my_apartment <- apartments_test[1, ]

lp_rf <- ceteris_paribus(explainer_rf, my_apartment)
head(lp_rf)

plot(lp_rf)

# --------
# multiclass

HR_rf <- randomForest(status ~ . , data = HR)
explainer_rf <- explain(HR_rf, data = HRTest, y = HRTest)

my_HR <- HRTest[1, ]

lp_rf <- ceteris_paribus(explainer_rf,  my_HR)
head(lp_rf)

plot(lp_rf, color = "_label_")

}
}
