% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_dependency.R
\name{partial_dependency}
\alias{partial_dependency}
\alias{partial_dependency.explainer}
\alias{partial_dependency.default}
\title{Partial Dependency Plots}
\usage{
partial_dependency(x, ...)

\method{partial_dependency}{explainer}(x, variables = NULL, N = 500,
  variable_splits = NULL, grid_points = 101, ...)

\method{partial_dependency}{default}(x, data, predict_function = predict,
  label = class(x)[1], variables = NULL, grid_points = grid_points,
  variable_splits = variable_splits, N = 500, ...)
}
\arguments{
\item{x}{a model to be explained, or an explainer created with function `DALEX::explain()`.}

\item{...}{other parameters}

\item{variables}{names of variables for which profiles shall be calculated. Will be passed to `calculate_variable_splits()`. If NULL then all variables from the validation data will be used.}

\item{N}{number of observations used for calculation of partial dependency profiles. By default 500.}

\item{variable_splits}{named list of splits for variables, in most cases created with `calculate_variable_splits()`. If NULL then it will be calculated based on validation data avaliable in the `explainer`.}

\item{grid_points}{number of points for profile. Will be passed to `calculate_variable_splits()`.}

\item{data}{validation dataset, will be extracted from `x` if it's an explainer}

\item{predict_function}{predict function, will be extracted from `x` if it's an explainer}

\item{label}{name of the model. By default it's extracted from the 'class' attribute of the model}
}
\value{
an 'aggregated_ceteris_paribus_explainer' layer
}
\description{
Function 'partial_dependency' calles 'ceteris_paribus' and then 'aggregate_profiles'
}
\examples{
library("DALEX")
 \dontrun{
library("titanic")
library("randomForest")

titanic_small <- titanic_train[,c("Survived", "Pclass", "Sex", "Age",
                                   "SibSp", "Parch", "Fare", "Embarked")]
titanic_small$Survived <- factor(titanic_small$Survived)
titanic_small$Sex <- factor(titanic_small$Sex)
titanic_small$Embarked <- factor(titanic_small$Embarked)
titanic_small <- na.omit(titanic_small)
rf_model <- randomForest(Survived ~ Pclass + Sex + Age + SibSp + Parch + Fare + Embarked,
                         data = titanic_small)
explainer_rf <- explain(rf_model, data = titanic_small,
                        y = titanic_small$Survived == "1", label = "RF")

pdp_rf <- partial_dependency(explainer_rf, variables = "Age")
pdp_rf

plot(pdp_rf)

pdp_rf <- partial_dependency(explainer_rf)
pdp_rf

plot(pdp_rf)
}
}
