% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspect_importance.R
\name{plot.aspect_importance}
\alias{plot.aspect_importance}
\title{Function for plotting aspect_importance results}
\usage{
\method{plot}{aspect_importance}(x, bar_width = 10, ...)
}
\arguments{
\item{x}{object of aspect_importance class}

\item{bar_width}{bar width}

\item{...}{other parameters}
}
\value{
a ggplot2 object
}
\description{
This function plots the results of aspect_importance.
}
