% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotD3.R
\name{plotD3}
\alias{plotD3}
\alias{plotD3.feature_importance_explainer}
\title{Plot Feature Importance Objects in D3 with r2d3 package.}
\usage{
plotD3(x, ...)

\method{plotD3}{feature_importance_explainer}(x, ..., label = TRUE,
  scaleHeight = FALSE)
}
\arguments{
\item{x}{the model model of `feature_importance_explainer` class.}

\item{...}{other parameters.}

\item{label}{default TRUE}

\item{scaleHeight}{default FALSE}
}
\value{
an `r2d3` object.
}
\description{
Plot Feature Importance Objects in D3 with r2d3 package.
}
\examples{
\dontrun{
## Not run:
set.seed(59)
library("DALEX")
library("caret")

regr_rf <- train(m2.price~., data = apartments, method="rf", ntree = 100)
regr_svm <- train(m2.price~., data = apartments, method="svmLinear")
explainer_rf <- explain(regr_rf, data = apartmentsTest[,2:6],
y = apartmentsTest$m2.price, label="rf")
explainer_svm <- explain(regr_svm, data = apartmentsTest[,2:6],
y = apartmentsTest$m2.price, label="svm")
fi_rf <- feature_importance(explainer_rf, loss_function = loss_root_mean_square)
fi_svm <- feature_importance(explainer_svm, loss_function = loss_root_mean_square)

plotD3(fi_rf, fi_svm)
plotD3(fi_rf, fi_svm, scaleHeight = TRUE)
plotD3(fi_rf, fi_svm, label = FALSE)
plotD3(fi_rf, fi_svm, label = FALSE, scaleHeight = TRUE)
}
}
