% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotD3_aggregated_profiles_explainer.R
\name{plotD3.aggregated_profiles_explainer}
\alias{plotD3.aggregated_profiles_explainer}
\title{Plots Aggregated Ceteris Paribus Profiles in D3 with r2d3 Package.}
\usage{
\method{plotD3}{aggregated_profiles_explainer}(x, ..., size = 2,
  alpha = 1, color = "#46bac2", only_numerical = TRUE,
  facet_ncol = 2, scale_plot = FALSE, variables = NULL,
  chartTitle = NULL, label_margin = 60)
}
\arguments{
\item{x}{a aggregated profiles explainer produced with function `aggregate_profiles()`}

\item{...}{other explainers that shall be plotted together}

\item{size}{a numeric. Set width of lines}

\item{alpha}{a numeric between 0 and 1. Opacity of lines}

\item{color}{a character.  Set line/bar color}

\item{only_numerical}{a logical. If TRUE then only numerical variables will be plotted. If FALSE then only selected variables will be plotted as bars.}

\item{facet_ncol}{number of columns for the `facet_wrap`}

\item{scale_plot}{a logical. Should size of the plot scale with window size? By default it's FALSE}

\item{variables}{if not NULL then only `variables` will be presented}

\item{chartTitle}{a character. Set custom title}

\item{label_margin}{a numeric. Set width of label margins, when only_numerical is FALSE}
}
\value{
a `r2d3` object.
}
\description{
Function 'plotD3.aggregated_profiles_explainer' plots an aggregate of ceteris paribus profiles.
It works in a similar way to 'plotD3.ceteris_paribus' but, instead of individual profiles,
show average profiles for each variable listed in the 'variables' vector.

Find more detailes in \href{https://pbiecek.github.io/PM_VEE/ceterisParibus.html}{Ceteris Paribus Chapter}.
}
\examples{
library("DALEX")
library("ingredients")
library("randomForest")
titanic <- na.omit(titanic)
model_titanic_rf <- randomForest(survived == "yes" ~ gender + age + class + embarked +
                                   fare + sibsp + parch,  data = titanic)

explain_titanic_rf <- explain(model_titanic_rf,
                              data = titanic[,-9],
                              label = "Random Forest v7")

selected_passangers <- select_sample(titanic, n = 100)
cp_rf <- ceteris_paribus(explain_titanic_rf, selected_passangers)

pdp_rf_p <- aggregate_profiles(cp_rf, type = "partial", only_numerical = TRUE)
pdp_rf_p$`_label_` <- "RF_partial"
pdp_rf_c <- aggregate_profiles(cp_rf, type = "conditional", only_numerical = TRUE)
pdp_rf_c$`_label_` <- "RF_conditional"
pdp_rf_a <- aggregate_profiles(cp_rf, type = "accumulated", only_numerical = TRUE)
pdp_rf_a$`_label_` <- "RF_accumulated"

plotD3(pdp_rf_p, pdp_rf_c, pdp_rf_a, only_numerical = TRUE)
plot(pdp_rf_p, pdp_rf_c, pdp_rf_a)

pdp <- aggregate_profiles(cp_rf, type = "partial", only_numerical = FALSE)
pdp$`_label_` <- "RF_partial"

plotD3(pdp, only_numerical = FALSE)

}
\references{
Predictive Models: Visual Exploration, Explanation and Debugging \url{https://pbiecek.github.io/PM_VEE}
}
