% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_aggregated_profiles.R
\name{print.aggregated_ceteris_paribus_explainer}
\alias{print.aggregated_ceteris_paribus_explainer}
\title{Prints Aggregated Profiles}
\usage{
\method{print}{aggregated_ceteris_paribus_explainer}(x, ...)
}
\arguments{
\item{x}{an individual variable profile explainer produced with the `aggregate_profiles()` function}

\item{...}{other arguments that will be passed to `head()`}
}
\description{
Prints Aggregated Profiles
}
\examples{
library("DALEX2")
 \dontrun{
}
}
