% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ceteris_paribus_descriptions.R
\name{print.ceteris_paribus_descriptions}
\alias{print.ceteris_paribus_descriptions}
\title{Prints Ceteris Paribus Descriptions}
\usage{
\method{print}{ceteris_paribus_descriptions}(x, ...)
}
\arguments{
\item{x}{an individual variable profile explainer produced with the `describe()` function}

\item{...}{other arguments}
}
\description{
Prints Ceteris Paribus Descriptions
}
\examples{
library("DALEX")
 \donttest{
}
}
