% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_neighbours.R
\name{select_neighbours}
\alias{select_neighbours}
\title{Select Subset of Rows Closest to a Specified Observation}
\usage{
select_neighbours(
  observation,
  data,
  variables = NULL,
  distance = gower::gower_dist,
  n = 20,
  frac = NULL
)
}
\arguments{
\item{observation}{single observation}

\item{data}{set of observations}

\item{variables}{names of variables that shall be used for calculation of distance.
By default these are all variables present in \code{data} and \code{observation}}

\item{distance}{the distance function, by default the \code{gower_dist()} function.}

\item{n}{number of neighbors to select}

\item{frac}{if \code{n} is not specified (NULL), then will be calculated as \code{frac} * number of rows in \code{data}.
Either \code{n} or \code{frac} need to be specified.}
}
\value{
a data frame with selected rows
}
\description{
Function \code{\link{select_neighbours}} selects subset of rows from data set.
This is useful if data is large and we need just a sample to calculate profiles.
}
\details{
Note that \code{select_neighbours()} function is S3 generic.
If you want to work on non standard data sources (like H2O ddf, external databases)
you should overload it.
}
\examples{
library("ingredients")

new_apartment <- DALEX::apartments[1,]
small_apartments <- select_neighbours(new_apartment, DALEX::apartments_test, n = 10)

new_apartment
small_apartments

}
