% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_aggregated_profiles.R
\name{show_aggreagated_profiles}
\alias{show_aggreagated_profiles}
\title{Adds a Layer with Aggregated Profiles}
\usage{
show_aggreagated_profiles(x, ..., size = 0.5, alpha = 1,
  color = "black", selected_variables = NULL)
}
\arguments{
\item{x}{a ceteris paribus explainer produced with function `ceteris_paribus()`}

\item{...}{other explainers that shall be plotted together}

\item{size}{a numeric. Size of lines to be plotted}

\item{alpha}{a numeric between 0 and 1. Opacity of lines}

\item{color}{a character. Either name of a color or name of a variable that should be used for coloring}

\item{selected_variables}{if not NULL then only `selected_variables` will be presented}

\item{sides}{a string containing any of "trbl", for top, right, bottom, and left. Passed to geom rug.}
}
\value{
a ggplot2 layer
}
\description{
Function 'show_aggreagated_profiles' adds a layer to a plot created with 'plot.ceteris_paribus_explainer'.
}
