% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_rugs.R
\name{show_rugs}
\alias{show_rugs}
\title{Adds a Layer with Rugs to a Profile Plot}
\usage{
show_rugs(x, ..., size = 0.5, alpha = 1, color = "#371ea3",
  only_numerical = TRUE, sides = "b", variables = NULL)
}
\arguments{
\item{x}{a ceteris paribus explainer produced with function `ceteris_paribus()`}

\item{...}{other explainers that shall be plotted together}

\item{size}{a numeric. Size of lines to be plotted}

\item{alpha}{a numeric between 0 and 1. Opacity of lines}

\item{color}{a character. Either name of a color or name of a variable that should be used for coloring}

\item{only_numerical}{a logical. If TRUE then only numerical variables will be plotted. If FALSE then only categorical variables will be plotted.}

\item{sides}{a string containing any of "trbl", for top, right, bottom, and left. Passed to geom rug.}

\item{variables}{if not NULL then only `variables` will be presented}
}
\value{
a ggplot2 layer
}
\description{
Function 'show_rugs' adds a layer to a plot created with 'plot.ceteris_paribus_explainer' for selected observations.
Various parameters help to decide what should be plotted, profiles, aggregated profiles, points or rugs.
}
\examples{
library("DALEX")
 \dontrun{
 library("randomForest")
 titanic <- na.omit(titanic)
 model_titanic_rf <- randomForest(survived == "yes" ~ gender + age + class + embarked +
                                    fare + sibsp + parch,  data = titanic)
 model_titanic_rf

 explain_titanic_rf <- explain(model_titanic_rf,
                           data = titanic[,-9],
                           y = titanic$survived == "yes",
                           label = "Random Forest v7")

selected_passangers <- select_sample(titanic, n = 100)
cp_rf <- ceteris_paribus(explain_titanic_rf , selected_passangers)
cp_rf

plot(cp_rf, variables = "Age") +
  show_observations(cp_rf, variables = "Age") +
  show_rugs(cp_rf, variables = "Age", color = "red")

}
}
