% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspect_importance.R
\name{triplot}
\alias{triplot}
\alias{triplot.explainer}
\alias{triplot.default}
\title{Three plots that sum up automatic aspect importance grouping}
\usage{
triplot(x, ...)

\method{triplot}{explainer}(x, new_observation, N = 500,
  clust_method = "complete", absolute_value = FALSE,
  cumulative_max = FALSE, ...)

\method{triplot}{default}(x, data, predict_function = predict,
  new_observation, N = 500, clust_method = "complete",
  absolute_value = FALSE, cumulative_max = FALSE, ...)
}
\arguments{
\item{x}{an explainer created with the \code{DALEX::explain()} function
or a model to be explained.}

\item{...}{other parameters}

\item{new_observation}{selected observation with columns that corresponds to
variables used in the model, should be without target variable}

\item{N}{number of rows to be sampled from data}

\item{clust_method}{the agglomeration method to be used, see
\code{\link[stats]{hclust}} methods}

\item{absolute_value}{if TRUE, aspect importance values will be drawn as
absolute values}

\item{cumulative_max}{if TRUE, aspect importance shown on tree will be max value
of children and node aspect importance values}

\item{data}{dataset, it will be extracted from \code{x} if it's an explainer
NOTE: Target variable shouldn't be present in the \code{data}}

\item{predict_function}{predict function, it will be extracted from \code{x} if
it's an explainer}
}
\description{
This function shows: \itemize{ \item plot for aspect_importance with
single aspect \item tree that shows aspect_importance for every newly
expanded aspect
\item clustering tree. }
}
\examples{
library(DALEX)
apartments_num <- apartments[,unlist(lapply(apartments, is.numeric))]
apartments_num_lm_model <- lm(m2.price ~ ., data = apartments_num)
apartments_num_new_observation <- apartments_num[2,-1]
apartments_num_mod <- apartments_num[,-1]
triplot(x = apartments_num_lm_model,
data = apartments_num_mod, new_observation = apartments_num_new_observation)


}
