% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspect_importance.R
\name{add_additional_information}
\alias{add_additional_information}
\title{temporary function, probably to be merged with aspect_importance}
\usage{
add_additional_information(x, data, aspect_list, show_cor = FALSE)
}
\arguments{
\item{x}{- aspect_importance() function results}

\item{data}{-  orignal data that was fed to aspect_importance() function}

\item{aspect_list}{- aspect list that was fed to aspect_importance() function}

\item{show_cor}{- binary parameter for showing or hiding information about correlations}
}
\value{
data frame
}
\description{
Add additional information to results of aspect_importance function
Show variables included in every aspect, minimal pairwise correlation in feature group and informs if any pairwise correlation in group is of negative type
}
