% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreagate_profiles.R
\name{aggregate_profiles}
\alias{aggregate_profiles}
\title{Aggregate Ceteris Paribus into Partial Dependency Plots}
\usage{
aggregate_profiles(x, ..., aggregate_function = mean,
  only_numerical = TRUE, selected_variables = NULL)
}
\arguments{
\item{x}{a ceteris paribus explainer produced with function `ceteris_paribus()`}

\item{...}{other explainers that shall be plotted together}

\item{aggregate_function}{a function for profile aggregation. By default it's 'mean'}

\item{only_numerical}{a logical. If TRUE then only numerical variables will be plotted. If FALSE then only categorical variables will be plotted.}

\item{selected_variables}{if not NULL then only `selected_variables` will be presented}
}
\value{
a 'aggregated_ceteris_paribus_explainer' layer
}
\description{
Function 'aggregate_profiles' calculate aggregate of ceteris paribus profiles
}
