% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspect_importance.R
\name{aspect_importance}
\alias{aspect_importance}
\alias{aspect_importance.explainer}
\alias{aspect_importance.default}
\alias{plot.aspect_importance}
\alias{aspect_lime}
\title{Aspect Importance
This explainer calculates the feature groups importance (called aspects importance) for a selected observation.}
\usage{
aspect_importance(x, ...)

\method{aspect_importance}{explainer}(x, new_observation, aspects_list,
  B = 100, method = "default")

\method{aspect_importance}{default}(model, data,
  predict_function = predict, new_observation, aspects_list, B = 100,
  method = "default")

\method{plot}{aspect_importance}(x, bar_width = 10)

aspect_lime(x, ...)
}
\arguments{
\item{x}{an explainer created with the `DALEX::explain()` function}

\item{new_observation}{selected observation with columns that corresponds to variables used in the model}

\item{aspects_list}{list containg grouping of features into aspects}

\item{B}{number of rows to be sampled from data}

\item{method}{sampling method in get_sample()}

\item{model}{model created on data, it will be extracted from `x` if it's an explainer}

\item{data}{dataset, it will be extracted from `x` if it's an explainer}

\item{predict_function}{predict function, it will be extracted from `x` if it's an explainer}

\item{bar_width}{bar width}

\item{x}{object of aspect_importance class}
}
\value{
An object of the class 'aspect_importance'.
Contains dataframe that describes aspects' importance.

a ggplot2 object
}
\description{
Aspect Importance function takes a sample from a given dataset and modifies it.
Modification is made by replacing part of its aspects by values from the observation.
Then function is calculating the difference between the prediction made on modified sample and the original sample.
Finally, it measures the impact of aspects on the change of prediction by using the linear model.

This function plots the results of aspect_importance
}
\examples{
library("DALEX")
titanic <- na.omit(titanic)
model_titanic_glm <- glm(survived == "yes" ~
                           class+gender+age+sibsp+parch+fare+embarked,
                         data = titanic, family = "binomial")

explain_titanic_glm <- explain(model_titanic_glm,
                               data = titanic[,-9],
                               y = titanic$survived == "yes")

aspects <- list(wealth = c("class", "fare"),
                family = c("gender", "sibsp", "parch"),
                age = "age",
                embarked = "embarked")

aspect_importance(explain_titanic_glm, new_observation = titanic[1,],
                  aspects_list = aspects)

\donttest{

library("randomForest")
model_titanic_rf <- randomForest(survived ~ class + gender + age + sibsp +
                                   parch + fare + embarked, data = titanic)

explain_titanic_rf <- explain(model_titanic_rf,
                              data = titanic[,-9],
                              y = titanic$survived == "yes")

aspect_importance(explain_titanic_rf, new_observation = titanic[1,],
                  aspects_list = aspects)

}

}
