% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_plots.R
\name{bind_plots}
\alias{bind_plots}
\title{Bind Multiple ggplot Objects}
\usage{
bind_plots(..., byrow = FALSE)
}
\arguments{
\item{...}{(`ggplot`) ggplot objects to combine.}

\item{byrow}{(`logical`) if FALSE (the default) the plots are bind by
columns, otherwise the plots are bind by rows.}
}
\value{
(`gtable`) A plottable object with \code{plot()}.
}
\description{
This is an aesthetically efficient implementation of the
  \link[gridExtra]{grid.arrange}.
}
\examples{
\dontrun{
mtcars_ggplot_0_0 <- ggplot2::ggplot(mtcars)
mtcars_ggplot_1_1 <- mtcars_ggplot_0_0 + ggplot2::geom_point(aes(x = mpg, y = wt))
mtcars_ggplot_1_2 <- mtcars_ggplot_1_1 + ggplot2::facet_grid(.~vs)
# Bind plots by rows: the result is a plot with 2 columns and 2 rows
bind_plots(mtcars_ggplot_1_1, mtcars_ggplot_1_2, byrow = TRUE)
# Bind plots by columns: the result is a plot with 3 columns and 1 row
bind_plots(mtcars_ggplot_1_1, mtcars_ggplot_1_2, byrow = FALSE)
}

}
