% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_profiles.R
\name{cluster_profiles}
\alias{cluster_profiles}
\title{Cluster Ceteris Paribus}
\usage{
cluster_profiles(x, ..., aggregate_function = mean,
  only_numerical = TRUE, center = FALSE, k = 3,
  selected_variables = NULL)
}
\arguments{
\item{x}{a ceteris paribus explainer produced with function `ceteris_paribus()`}

\item{...}{other explainers that shall be plotted together}

\item{aggregate_function}{a function for profile aggregation. By default it's 'mean'}

\item{only_numerical}{a logical. If TRUE then only numerical variables will be plotted. If FALSE then only categorical variables will be plotted.}

\item{center}{shall profiles be centered before clustering}

\item{k}{number of clusters for the hclust function}

\item{selected_variables}{if not NULL then only `selected_variables` will be presented}
}
\value{
a 'aggregated_ceteris_paribus_explainer' layer
}
\description{
Function 'cluster_profiles' calculates aggregate of ceteris paribus profiles
}
\examples{
library("DALEX")
 \dontrun{
library("titanic")
library("randomForest")

titanic_small <- titanic_train[,c("Survived", "Pclass", "Sex", "Age",
                                   "SibSp", "Parch", "Fare", "Embarked")]
titanic_small$Survived <- factor(titanic_small$Survived)
titanic_small$Sex <- factor(titanic_small$Sex)
titanic_small$Embarked <- factor(titanic_small$Embarked)
titanic_small <- na.omit(titanic_small)
rf_model <- randomForest(Survived ~ Pclass + Sex + Age + SibSp + Parch + Fare + Embarked,
                         data = titanic_small)
predict_fuction <- function(m,x) predict(m, x, type="prob")[,2]
explainer_rf <- explain(rf_model, data = titanic_small,
                        y = titanic_small$Survived == "1", label = "RF",
                        predict_function = predict_rf_fuction)

selected_passangers <- select_sample(titanic_small, n = 100)
cp_rf <- ceteris_paribus(explainer_rf, selected_passangers)
cp_rf

pdp_rf <- aggregate_profiles(cp_rf, selected_variables = "Age")
pdp_rf
clust_rf <- cluster_profiles(cp_rf, k = 3, selected_variables = "Age")
clust_rf

plot(clust_rf, color = "_label_") +
  show_aggreagated_profiles(pdp_rf, color = "black", size = 3)

plot(cp_rf, color = "grey") +
  show_aggreagated_profiles(clust_rf, color = "_label_")

clust_rf <- cluster_profiles(cp_rf, k = 3, center = TRUE, selected_variables = "Age")
clust_rf

plot(clust_rf, color = "_label_") +
  show_aggreagated_profiles(pdp_rf, color = "black", size = 3)
}
}
