% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotD3_ceteris_paribus_explainer.R
\name{plotD3}
\alias{plotD3}
\alias{plotD3.ceteris_paribus_explainer}
\title{Plots Ceteris Paribus Profiles in D3 with r2d3 Package.}
\usage{
plotD3(x, ...)

\method{plotD3}{ceteris_paribus_explainer}(x, ..., size = 2, alpha = 1,
  color = "#46bac2", only_numerical = TRUE, facet_ncol = 2,
  scale_plot = FALSE, variables = NULL, chartTitle = NULL,
  label_margin = 60, show_observations = TRUE, show_rugs = FALSE)
}
\arguments{
\item{x}{a ceteris paribus explainer produced with function `ceteris_paribus()`}

\item{...}{other explainers that shall be plotted together}

\item{size}{a numeric. Set width of lines}

\item{alpha}{a numeric between 0 and 1. Opacity of lines}

\item{color}{a character.  Set line color}

\item{only_numerical}{a logical. If TRUE then only numerical variables will be plotted. If FALSE then only selected variables will be plotted as bars.}

\item{facet_ncol}{number of columns for the `facet_wrap`}

\item{scale_plot}{a logical. Should size of the plot scale with window size? By default it's FALSE}

\item{variables}{if not NULL then only `variables` will be presented}

\item{chartTitle}{a character. Set custom title}

\item{label_margin}{a numeric. Set width of label margins, when only_numerical is FALSE}

\item{show_observations}{a logcal. Adds observations layer to a plot. By default it's TRUE}

\item{show_rugs}{a logcal. Adds rugs layer to a plot. By default it's FALSE}
}
\value{
an `r2d3` object.
}
\description{
Function 'plotD3.ceteris_paribus_explainer' plots Individual Variable Profiles for selected observations.
Various parameters help to decide what should be plotted, profiles, aggregated profiles, points or rugs.
}
\details{
Find more detailes in \href{https://pbiecek.github.io/PM_VEE/ceterisParibus.html}{Ceteris Paribus Chapter}.
}
\examples{
\dontrun{
library("DALEX")
library("ingredients")
library("randomForest")

titanic <- na.omit(titanic)

model_titanic_rf <- randomForest(survived == "yes" ~ gender + age + class + embarked +
                                 fare + sibsp + parch,  data = titanic)
explain_titanic_rf <- explain(model_titanic_rf,
                             data = titanic[,-9],
                             y = titanic$survived == "yes",
                             label = "rf")

selected_passangers <- select_sample(titanic, n = 10)
cp_rf <- ceteris_paribus(explain_titanic_rf, selected_passangers)

plotD3(cp_rf, variables = c("age","parch","fare","sibsp"),
     facet_ncol = 2, scale_plot = TRUE)

selected_passanger <- select_sample(titanic, n = 1)
cp_rf <- ceteris_paribus(explain_titanic_rf, selected_passanger)

plotD3(cp_rf, variables = c("class", "embarked", "gender", "sibsp"),
     facet_ncol = 2, only_numerical = FALSE, label_margin = 100, scale_plot = TRUE)
}

}
