% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_aggregated_profiles.R
\name{print.aggregated_ceteris_paribus_explainer}
\alias{print.aggregated_ceteris_paribus_explainer}
\title{Prints Aggregated Profiles}
\usage{
\method{print}{aggregated_ceteris_paribus_explainer}(x, ...)
}
\arguments{
\item{x}{an individual variable profile explainer produced with the `aggregate_profiles()` function}

\item{...}{other arguments that will be passed to `head()`}
}
\description{
Prints Aggregated Profiles
}
\examples{
library("DALEX")
 \dontrun{
library("titanic")
library("randomForest")

titanic_small <- titanic_train[,c("Survived", "Pclass", "Sex", "Age", "SibSp", "Parch", "Fare", "Embarked")]
titanic_small$Survived <- factor(titanic_small$Survived)
titanic_small$Sex <- factor(titanic_small$Sex)
titanic_small$Embarked <- factor(titanic_small$Embarked)
titanic_small <- na.omit(titanic_small)
rf_model <- randomForest(Survived ~ Pclass + Sex + Age + SibSp + Parch + Fare + Embarked,
                         data = titanic_small)
predict_fuction <- function(m,x) predict(m, x, type="prob")[,2]
explainer_rf <- explain(rf_model, data = titanic_small,
                        y = titanic_small$Survived == "1", label = "RF",
                        predict_function = predict_rf_fuction)

selected_passangers <- select_sample(titanic_small, n = 100)
cp_rf <- ceteris_paribus(explainer_rf, selected_passangers)
cp_rf

pdp_rf <- aggregate_profiles(cp_rf, selected_variables = "Age")
pdp_rf

}
}
