% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_aggregated_profiles.R
\name{show_aggreagated_profiles}
\alias{show_aggreagated_profiles}
\title{Adds a Layer with Aggregated Profiles}
\usage{
show_aggreagated_profiles(x, ..., size = 0.5, alpha = 1,
  color = "black", selected_variables = NULL)
}
\arguments{
\item{x}{a ceteris paribus explainer produced with function `ceteris_paribus()`}

\item{...}{other explainers that shall be plotted together}

\item{size}{a numeric. Size of lines to be plotted}

\item{alpha}{a numeric between 0 and 1. Opacity of lines}

\item{color}{a character. Either name of a color or name of a variable that should be used for coloring}

\item{selected_variables}{if not NULL then only `selected_variables` will be presented}
}
\value{
a ggplot2 layer
}
\description{
Function 'show_aggreagated_profiles' adds a layer to a plot created with 'plot.ceteris_paribus_explainer'.
}
\examples{
library("DALEX")
 \dontrun{
library("titanic")
library("randomForest")

titanic_small <- titanic_train[,c("Survived", "Pclass", "Sex", "Age",
                                   "SibSp", "Parch", "Fare", "Embarked")]
titanic_small$Survived <- factor(titanic_small$Survived)
titanic_small$Sex <- factor(titanic_small$Sex)
titanic_small$Embarked <- factor(titanic_small$Embarked)
titanic_small <- na.omit(titanic_small)
rf_model <- randomForest(Survived ~ Pclass + Sex + Age + SibSp + Parch + Fare + Embarked,
                         data = titanic_small)
predict_fuction <- function(m,x) predict(m, x, type="prob")[,2]
explainer_rf <- explain(rf_model, data = titanic_small,
                        y = titanic_small$Survived == "1", label = "RF",
                        predict_function = predict_rf_fuction)

selected_passangers <- select_sample(titanic_small, n = 100)
cp_rf <- ceteris_paribus(explainer_rf, selected_passangers)
cp_rf

pdp_rf <- aggregate_profiles(cp_rf, selected_variables = "Age")
pdp_rf

plot(cp_rf, selected_variables = "Age", color = "grey") +
show_observations(cp_rf, selected_variables = "Age", color = "grey") +
  show_rugs(cp_rf, selected_variables = "Age", color = "red") +
  show_aggreagated_profiles(pdp_rf, size = 3)

plot(pdp_rf, selected_variables = "Age", color = "grey")

}
}
