% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_observations.R
\name{show_observations}
\alias{show_observations}
\title{Adds a Layer with Observations to a Profile Plot}
\usage{
show_observations(x, ..., size = 2, alpha = 1, color = "black",
  only_numerical = TRUE, selected_variables = NULL)
}
\arguments{
\item{x}{a ceteris paribus explainer produced with function `ceteris_paribus()`}

\item{...}{other explainers that shall be plotted together}

\item{size}{a numeric. Size of lines to be plotted}

\item{alpha}{a numeric between 0 and 1. Opacity of lines}

\item{color}{a character. Either name of a color or name of a variable that should be used for coloring}

\item{only_numerical}{a logical. If TRUE then only numerical variables will be plotted. If FALSE then only categorical variables will be plotted.}

\item{selected_variables}{if not NULL then only `selected_variables` will be presented}
}
\value{
a ggplot2 layer
}
\description{
Function 'show_observations' adds a layer to a plot created with 'plot.ceteris_paribus_explainer' for selected observations.
Various parameters help to decide what should be plotted, profiles, aggregated profiles, points or rugs.
}
